requirements.txt
docling>=2.0
pgvector>=0.2.0
psycopg2-binary>=2.9
openai>=1.0
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pandas>=2.0
bcrypt>=4.0.0
tenacity>=8.0.0
pydantic-settings>=2.0.0
